PropVals can now set Prop mgrs
Added applyProperty(PropertyMgr*) method to the PropertyVal objects so
that they can set the appropriate property of a tool with the desired
type. This introduces a dependency on PropertyMgr in the including code.

A nicer approach could be possible where we use a getProperty method to
return a new asg::Property object which can be pushed into the
PropertyMgr in some client code. This, however, requires an update to
the PropertyMgr interface in RootCore.